- updated AbstractFurnaceBlock - added Register - updated json_builder - increased smoke spawn rate - added DebugController (F2 + T to open test interface) - added interface opening animation - removed method of registering a string ID for a specific numeric ID - fixed a crash when adding an item to a hotbar - code refactoring

+++ core/src/com/kgc/sauw/core/environment/block/HeatingBlock.java
@@ -3,5 +3,5 @@ package com.kgc.sauw.core.environment.block;
 import com.kgc.sauw.core.environment.world.Tile;
 
 public interface HeatingBlock {
-    float heatTemperature(Tile tile);
+    double heatTemperature(Tile tile);
 }
